refactor out create_file_if_not_exists
